+++ join/h1.hpp
@@ -1,10 +1,10 @@
 #include "structs.hpp"
 
-int32_t h1(int32_t value);
-uint32_t * calculateHistogram(Relation * rel);
-uint32_t * calculateStartingPositions(uint32_t * histogram);
-void printHistogram(uint32_t * histogram);
-Tuple * order(Relation * rel, uint32_t * startingPositions);
+uint64_t h1(uint64_t value);
+uint64_t * calculateHistogram(Relation * rel);
+uint64_t * calculateStartingPositions(uint64_t * histogram);
+void printHistogram(uint64_t * histogram);
+Tuple * order(Relation * rel, uint64_t * startingPositions);
 Tuple * bucketify(Relation * rel,
-                  uint32_t ** histogram,
-                  uint32_t ** startingPositions);
+                  uint64_t ** histogram,
+                  uint64_t ** startingPositions);
